Begin Content Creation
Content can be created by calling the newPost method and providing the ID of the user creating the post, the body of the post, and an optional file to submit.

Missing features include linking to listing, liking features etc.

There is also now a helper method to fetch user information by username. The ID of the user is required to be passed into each new content creation request or an error will be thrown.

Please use service_role API key instead of anon key.