Added post removal & getPostByID
Post deletion can be done with method removePost(user_id, post_id). The user ID can be found by using the findUserID query with the string correlating to the username. post_id is found in the return from getPosts.

Also added a getPostsByID method where a specific post can be returned if you enter the post_id. An empty array is returned if there is no post with that ID.